requirements.txt
# Core deep learning
torch>=2.1.0
torchvision>=0.16.0

# ViT building blocks
timm>=0.9.0

# Config
pyyaml>=6.0

# Data
h5py>=3.9.0          # PatchCamelyon HDF5 files
Pillow>=10.0.0
numpy>=1.24.0

# WSI processing (optional, required for CAMELYON16 / TCGA preprocessing)
openslide-python>=1.3.0
opencv-python>=4.8.0

# Experiment tracking (optional)
tensorboard>=2.14.0
wandb>=0.16.0

# Evaluation & visualisation (optional)
scikit-learn>=1.3.0   # t-SNE
matplotlib>=3.7.0
umap-learn>=0.5.4
